requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
django-environ==0.11.2
fastapi==0.109.2
graphql-core==3.2.3
h11==0.14.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.2.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
strawberry-graphql==0.219.2
threadpoolctl==3.3.0
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
